Consolidate FTS writes through SearchProjectionService
